Don't allow negative console numbers for the VT_WAITACTIVE ioctl.  Use the correct console number for the VT_WAITACTIVE ioctl.  Invalid console numbers caused waiting on an invalid pointer.
Use bcopyw() instead of move_up() and move_down().  bcopyw() handles
overlapped copies and should be faster.  Actually use bcopy().  bcopy()
is slightly faster if video memory is 16-bit and about twice as fast if
it is 32-bit.  bcopy() is said to fail on someGA's, but syscons already
depends on it working for other accesses to video memory.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@5359 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f