Code to prevent a panic caused by procfs_exit().  Note that i don't know what is teh root cause -- but, sometimes, a procfs vnode in pfshead is apparantly corrupt (or a UFS vnode instead).  Without this patch, I can get it to panic by doing (in csh)
	while (1)
		ps auxwww
	end

and it will panic when the PID's wrap.  With it, it does not panic.
Yes -- I know that this is NOT the right way to fix it.  But I haven't
been able to get it to panic yet (which confuses me).  I am going to
be looking into the vgone() code now, as that may be a part of it.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@31640 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f