Shoot the LKM support in the old wd/wdc/atapi driver set in the head and perform a cleanup/unifdef sweep over it to tidy things up.  The atapi code is permanently attached to the wd driver and is always probed.
I will add an extra option bit in the flags to disable an atapi probe on
either the master or slave if needed, if people want this.

Remember, this driver is destined to die some time.  It's possible that
it will loose all atapi support down the track and only be used for
dumb non-ATA disks and all ata/atapi devices will be handled by the new
ata system.

ATAPI, ATAPI_STATIC and CMD640 are no longer options, all are implicit.

Previously discussed with:  sos


git-svn-id: svn+ssh://svn.freebsd.org/base/head@45666 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f