Install sendmail.cf into /etc/mail.
The current offical Sendmail Inc. version uses /var/mail/ and when we upgrade
our repository to that version, we will get the change.  It is best to make
the path change in 4.0-R (which may not have the latest Sendmail Inc. version,
than to change in mid-4.x stream when we may upgrade.

Ok'ed by:	Peter (quite a while ago)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@54843 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f