Turn a warning into an error. Occurances of that warning generate bad -fpic code that damages symbol locations at runtime.
The only know occurance in our tree (src and ports) was locatime.c,
which was just changed to generate code that doesn't trigger the
problem.

This is a workaround, the real cause is that our gas doesn't
understand code our gcc generates for some -O -fpic code. They are
expected to be back in sync soon, but until then (including
4.0-RELEASE) we need to prevent people from using bad -fpic code.

PR:		avoids such things as in bin/16862
Submitted by:	bde
Approved by:	jkh


git-svn-id: svn+ssh://svn.freebsd.org/base/head@57823 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f